Last push before submitting the project for milestone 3
Current issues
Big: The instruction set doesn't support read or write to or from user. Have not figured out a solution for allowing the model to comunicate to the view that it wants input or output. Breaks MVVM model

Not all controls have functionality behind them. Particularly the File toolbar menue items. However most necessary functionality is currently implemented in the flyout context menues

Have seen in debugging mode that the program will sometimes launch as a solid white screen. Needs to be relaunched in such cases